Refactor: group dashboard viz routes
