+++ packages/admin-panel-server/src/routes/dashboardVisualisations/index.ts
@@ -1,0 +1,10 @@
+/*
+ * Tupaia
+ * Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
+ *
+ */
+
+export * from './ExportDashboardVisualisationRoute';
+export * from './FetchDashboardVisualisationRoute';
+export * from './ImportDashboardVisualisationRoute';
+export * from './SaveDashboardVisualisationRoute';
